fix https://github.com/microsoft/vscode/issues/194414
