+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "web-ai-toolkit",
-  "version": "0.1.8",
+  "version": "0.2.0",
   "repository": "https://github.com/jgw96/web-ai-toolkit",
   "keywords": [
     "ai",
@@ -29,12 +29,12 @@
   "author": "",
   "license": "ISC",
   "devDependencies": {
-    "typescript": "^5.5.3",
-    "vite": "^5.3.3",
-    "vite-plugin-dts": "^3.9.1"
+    "typescript": "^5.6.2",
+    "vite": "^5.4.8",
+    "vite-plugin-dts": "^4.2.2"
   },
   "dependencies": {
-    "@huggingface/transformers": "^3.0.0-alpha.14",
+    "@huggingface/transformers": "^3.0.0-alpha.16",
     "@xenova/transformers": "^2.17.2"
   }
 }
